+++ changelog.md
@@ -1,8 +1,13 @@
-# Version 1.40.2 (2023-08-08)
+# Version 1.40.2 (2023-08-12)
 
+- Bump hahomematic to 2023.8.3
+  - Update project setup
+  - Restructure test helper
+  - Add ping pong tests
 - Improve custom component testing
 - Increase config flow coverage
 - Add Github flows for pylint and tests
+- Add infrastructure for platform tests
 
 # Version 1.40.1 (2023-08-06)